+++ TDS/Assets/2.Scripts/Zombie/ZombieModel.cs
@@ -6,23 +6,23 @@ namespace TDS.Zombie
     public class ZombieModel
     {
         // 기본 스탯
-        private readonly ReactiveProperty<float> _maxHp;
-        private readonly ReactiveProperty<float> _currentHp;
+        private readonly ReactiveProperty<int> _maxHp;
+        private readonly ReactiveProperty<int> _currentHp;
         private readonly ReactiveProperty<bool> _isDead;
         
         // 읽기 전용 프로퍼티
-        public IReadOnlyReactiveProperty<float> MaxHp => _maxHp;
-        public IReadOnlyReactiveProperty<float> CurrentHp => _currentHp;
+        public IReadOnlyReactiveProperty<int> MaxHp => _maxHp;
+        public IReadOnlyReactiveProperty<int> CurrentHp => _currentHp;
         public IReadOnlyReactiveProperty<bool> IsDead => _isDead;
         
         // 현재 체력 비율 (0 ~ 1)
-        public float HpRatio => _currentHp.Value / _maxHp.Value;
+        public float HpRatio => (float)_currentHp.Value / _maxHp.Value;
         private CompositeDisposable disposables = new CompositeDisposable();
         
-        public ZombieModel(float maxHp)
+        public ZombieModel(int maxHp)
         {
-            _maxHp = new ReactiveProperty<float>(maxHp);
-            _currentHp = new ReactiveProperty<float>(maxHp);
+            _maxHp = new ReactiveProperty<int>(maxHp);
+            _currentHp = new ReactiveProperty<int>(maxHp);
             _isDead = new ReactiveProperty<bool>(false);
             
             // 체력이 0 이하가 되면 사망 상태로 전환
@@ -32,7 +32,7 @@ namespace TDS.Zombie
                 .AddTo(disposables);
         }
         
-        public void TakeDamage(float damage)
+        public void TakeDamage(int damage)
         {
             if (_isDead.Value) return;
